pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools"]

[project]
name = "recordcollection"
authors = [
    {name = "Robert Huselius", email = "robert@huseli.us"},
]
description = "Record collection"
license = {file = "LICENSE"}
requires-python = ">=3.10"
dependencies = [
    "django",
    "python-dotenv",
    "requests",
    "django-extensions",
    "tinytag",
    "levenshtein",
]
dynamic = ["version"]

[project.optional-dependencies]
dev = [
    "ipdb",
    "ipython",
    "isort",
    "pylint",
    "flake8",
    "pylint-django",
    "types-requests",
    "django-debug-toolbar",
]

[tool.setuptools.dynamic]
version = {attr = "recordcollection.__version__"}

[tool.pylint.main]
disable = [
    "import-outside-toplevel",
    "missing-class-docstring",
    "missing-function-docstring",
    "missing-module-docstring",
    "broad-exception-caught",
    "too-few-public-methods",
    "too-many-branches",
    "too-many-instance-attributes",
    "invalid-str-returned",
    "too-many-locals",
    "too-many-statements",
    "too-many-ancestors",
    "too-many-return-statements",
]
ignore = [".venv"]
max-line-length = 119

[tool.isort]
combine_as_imports = true
default_section = "THIRDPARTY"
include_trailing_comma = true
line_length = 79
multi_line_output = 3
lines_after_imports = 2
sections = [
    "FUTURE",
    "STDLIB",
    "THIRDPARTY",
    "FIRSTPARTY",
    "LOCALFOLDER",
]
skip = [".venv"]
